Filter scoped modules correctly
Fixes microsoft/license-checker-webpack-plugin#25

A valid node module name is either `\w+` or `@\w+/\w+` never `@\w+`.

This is enough of a hairy internal implementation detail that filter
should probably not be exposed even as an (undocumented) option. But
leaving alone for now as changing the pattern has been a useful
workaround.